add PaymentMethodCard component
